+++ CHANGELOG.md
@@ -4,7 +4,14 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.9.24 - Unreleased
+## Version 0.9.25 - Unreleased
+
+### Fixed
+
+* `kwimage.Boxes.clip` now works on torch.Tensor data.
+
+
+## Version 0.9.24 - Released 2024-02-23
 
 ### Added